// src/components/content/video/d.ts/config.d.ts
export declare interface VideoContextInterface {
    storageItemKey: string;
    storageItemValue: string;
    videoStep: string,
    videoLikeNum: string;
    videoAnnounce: string;
}

export declare interface VideoFuncBarInterface {
    name: string;
    icon: string | Array<string>;
    word: string | Array<string>;
    clickFunc: string;
}